split-files: refactor (same as https://github.com/privacy-scaling-explorations/zkevm-circuits/pull/1129)
